requirements.txt
azure-ai-ml
azure-identity
mlflow
scikit-learn
joblib
numpy